+++ packages/artifact-tool/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @opengeni/artifact-tool
 
+## 0.3.26
+
+### Patch Changes
+
+- Updated dependencies [e1a50ba]
+  - @opengeni/contracts@3.0.2
+
 ## 0.3.25
 
 ### Patch Changes